Isolate loading logic
